requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
python-docx==0.8.11
PyPDF2==3.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
python-Levenshtein==0.21.1
Flask-Caching==2.0.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.0
pytest==7.3.1
gunicorn==20.1.0 